pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-intruder"
version = "0.1.0"
description = "LLM-Intruder — LLM Security Assessment Tool for authorised penetration testers"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "playwright>=1.40",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
dashboard = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "websockets>=12.0",
]

[project.scripts]
llm-intruder = "llm_intruder.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["llm_intruder"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.metadata]
allow-direct-references = true
